Refactor to util library
utilize_params_util.py now stores functions which may be used across all
scripts in this directory.
